Extract Cilium reconciler
